+++ components/common/ConnectWallet/WalletAssetIdentity.vue
@@ -9,10 +9,6 @@
       <a v-clipboard:copy="account" @click="toast('Copied to clipboard')">
         <NeoIcon icon="copy" />
       </a>
-      <hr class="my-1" />
-      <a @click="logout()">
-        <NeoIcon icon="right-from-bracket" />
-      </a>
     </div>
   </div>
 </template>
@@ -20,12 +16,8 @@
 <script setup lang="ts">
 import { NeoIcon } from '@kodadot1/brick'
 import { useIdentityStore } from '@/stores/identity'
-import { useShoppingCartStore } from '@/stores/shoppingCart'
-import { useWalletStore, walletHistory } from '@/stores/wallet'
 
 const identityStore = useIdentityStore()
-const shoppingCartStore = useShoppingCartStore()
-const walletStore = useWalletStore()
 const { urlPrefix } = usePrefix()
 const { toast } = useToast()
 
@@ -34,13 +26,4 @@ const account = computed(() => identityStore.getAuthAddress)
 const { display, shortenedAddress } = useIdentity({
   address: account,
 })
-
-const logout = () => {
-  identityStore.resetAuth()
-  sessionStorage.clear()
-  localStorage.clear()
-  shoppingCartStore.clear()
-
-  walletHistory.value = walletStore.history
-}
 </script>
